Fix `keyup` event listener in `Layout` component. (#1369)
* Fix `keyup` event listener in `Layout` component.

This fixes a typo on the `keyup` event listener in the `Layout`
component. Previously, this component was listening for `keyUp`, but the
event should be [`keyup`](https://developer.mozilla.org/en-US/docs/Web/API/Document/keyup_event).

This fixes an issue where ZenMode could be enabled by pressing `alt` and
then `z` separately instead of together
(https://github.com/zeit/docs/issues/1358). Because `onKeyUp` was not
being triggered, we never updated the state to indicate `alt` was no
longer being pressed down.

* Fix bug tracking whether `alt` is held down in `Layout` component.

Previously in the `onKeyUp` function in the `Layout` component,
`this.altKeyDown` would **always** end up being set to `false`, no
matter which key was being released. This change only sets
`this.altKeyDown` to `false` if `event.which` indicates that it was the
`alt` key (18) that was released (otherwise, it keeps the state as-is).

This resolves an issue where you could not hold `alt` and press `z`
multiple times to toggle ZenMode.