Change directory the build should be output to (#1549)
* Change directory the build should be output to

This seems to have changed to 'dist' where it was 'public' before? I ran into errors and changing this value (as per the error message) got it working again.

* Remove configuration parts of the Nuxt.js guide

Nuxt.js is now automatically detected, removing the need for extra configuration of the output directory. Even more so, the original configuration in the Nuxt.js guide now causes an error. If you remove the configuration, it starts working again.